Updated Go script for docker-compose-yml generation to utilize Cobra, and added Go vendoring.
Closes #11
