+++ src/go.mod
@@ -1,0 +1,10 @@
+module sonic-lab
+
+go 1.25.5
+
+require github.com/spf13/cobra v1.10.2
+
+require (
+	github.com/inconshreveable/mousetrap v1.1.0 // indirect
+	github.com/spf13/pflag v1.0.9 // indirect
+)
